pyproject.toml
[project]
name = "rag-assign"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.127.0",
    "langchain>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "supervisor>=4.3.0",
    "uvicorn>=0.40.0",
]
